demos/knowledge-base/requirements.txt
mem0ai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
